refactor(overlay): split mixed-responsibility files added by this PR
Three unrelated concerns had accreted into files owning a different
primary responsibility:

- chunkArgs/maxDpkgArgBytes (arg-list batching) lived in install.go,
  which owns the deb/rpm installer backends. Moved to chunk.go.
- ArtifactProvides + its deb/rpm readers lived in depcheck.go, which
  otherwise owns Depends/Obsoletes/Conflicts reading. Moved to
  provides.go.
- classifyIntraSetConflicts (to-install-vs-to-install conflict
  classification) lived in preflight.go alongside the two-slice
  baseline classifier. Moved to intraset_conflict.go.

Each accompanying test moved with its subject to the matching
*_test.go file. No behavior change; verified via gofmt, go vet, and
the full overlay package test suite.
